refactor(agents): split Claude lifecycle tests
